+++ source/staking/scripts/deploy.js
@@ -1,7 +1,9 @@
 /* eslint-disable no-console */
 const fs = require('fs')
+const Confirm = require('prompt-confirm')
 const { ethers, run } = require('hardhat')
 const { chainNames, stakingTokenAddresses } = require('@airswap/constants')
+const { getEtherscanURL } = require('@airswap/utils')
 const stakingDeploys = require('../deploys.js')
 
 async function main() {
@@ -10,6 +12,7 @@ async function main() {
   console.log(`Deployer: ${deployer.address}`)
 
   const chainId = await deployer.getChainId()
+  const gasPrice = await deployer.getGasPrice()
   const stakingToken = stakingTokenAddresses[chainId]
   const name = 'Staked AST'
   const symbol = 'sAST'
@@ -17,29 +20,39 @@ async function main() {
   const minDelay = 2419200
 
   console.log(`Deploying on ${chainNames[chainId].toUpperCase()}`)
-  const stakingFactory = await ethers.getContractFactory('Staking')
-  const stakingContract = await stakingFactory.deploy(
-    stakingToken,
-    name,
-    symbol,
-    duration,
-    minDelay
-  )
-  await stakingContract.deployed()
-  console.log(`Deployed: ${stakingContract.address}`)
+  console.log(`Staking token: ${stakingToken}`)
+  console.log(`Gas price: ${gasPrice / 10 ** 9} gwei`)
 
-  stakingDeploys[chainId] = stakingContract.address
-  fs.writeFileSync(
-    './deploys.js',
-    `module.exports = ${JSON.stringify(stakingDeploys, null, '\t')}`
-  )
-  console.log('Updated deploys.js')
+  const prompt = new Confirm('Proceed to deploy?')
+  if (await prompt.run()) {
+    const stakingFactory = await ethers.getContractFactory('Staking')
+    const stakingContract = await stakingFactory.deploy(
+      stakingToken,
+      name,
+      symbol,
+      duration,
+      minDelay
+    )
+    console.log(
+      'Deploying...',
+      getEtherscanURL(chainId, stakingContract.deployTransaction.hash)
+    )
+    await stakingContract.deployed()
+    console.log(`Deployed: ${stakingContract.address}`)
 
-  console.log(
-    `\nVerify with "yarn verify --network ${chainNames[
-      chainId
-    ].toLowerCase()}"\n`
-  )
+    stakingDeploys[chainId] = stakingContract.address
+    fs.writeFileSync(
+      './deploys.js',
+      `module.exports = ${JSON.stringify(stakingDeploys, null, '\t')}`
+    )
+    console.log('Updated deploys.js')
+
+    console.log(
+      `\nVerify with "yarn verify --network ${chainNames[
+        chainId
+      ].toLowerCase()}"\n`
+    )
+  }
 }
 
 main()
